+++ _src/core/struct.js
@@ -1,75 +1,3 @@
-import GObject from './gobject.js';
-
-class GString {
-    static #get_array(str) {
-        let strs = [''], i = 0, is_in_val = false, is_locating_quote = false, skip_a_char = false, quote_locate_starter = '{';
-        str.split("").filter(v => v != '').forEach((char) => {
-            if (!skip_a_char) {
-                if (char == '\\') {
-                    skip_a_char = true;
-                } else if (is_locating_quote) {
-                    if (char == '{' || char == '}') {
-                        is_in_val = !is_in_val;
-                        i++;
-                        strs[i] = '';
-                    } else {
-                        strs[i] += quote_locate_starter + char;
-                    }
-                    is_locating_quote = false;
-                } else if (char == '{' && !is_in_val) {
-                    is_locating_quote = true;
-                    quote_locate_starter = '{';
-                } else if (char == '}' && is_in_val) {
-                    is_locating_quote = true;
-                    quote_locate_starter = '}';
-                } else
-                    strs[i] += char;
-            } else {
-                strs[i] += char;
-                skip_a_char = false
-            }
-        });
-        return strs;
-    }
-
-    static #eval(arr, coll) {
-        if (!coll) return this.toString();
-        let is_in_val = false, result = '';
-        arr.forEach((v) => {
-            if (is_in_val) result += coll.get(v);
-            else result += v;
-            is_in_val = !is_in_val;
-        });
-        return result;
-    }
-
-    static from(str) {
-        let strs = GString.#get_array(str);
-        return new SymString(strs);
-    }
-
-    static parse(str, coll) {
-        return GString.#eval(GString.#get_array(str), coll);
-    }
-
-    #str_group;
-    constructor(strg) {
-        if (Array.isArray(strg)) {
-            this.#str_group = strg;
-        } else throw new Error('SymString constructor : Argument  not any[]');
-    }
-    /**
-     * 
-     * @param {Collection} collection 
-     */
-    eval(collection) {
-        return GString.#eval(collection);
-    }
-    toString() {
-        return this.#str_group.join();
-    }
-}
-
 class BuiltinCorrespond {
     from = '';
     to = '';
@@ -77,33 +5,6 @@ class BuiltinCorrespond {
         this.from = from;
         this.to = to;
     }
-}
-
-class Collection {
-    #obj;
-    /**
-     * 
-     * @param {object} obj 
-     */
-    constructor(obj) {
-        this.#obj = obj;
-    }
-    get(key) {
-        return GObject.getProperty(this.#obj, key);
-    }
-    has(key) {
-        return GObject.hasProperty(this.#obj, key);
-    }
-    append(ext) {
-        this.#obj = GObject.mix(this.#obj, ext, true);
-    }
-    asObject() {
-        return this.#obj;
-    }
-    /**
-     * @deprecated use asObject insteads
-     */
-    get_all = this.asObject;
 }
 
 class Cache {
@@ -138,8 +39,6 @@ function NULL_OBJECT() {
 }
 
 export {
-    GString,
-    Collection,
     Cache,
     Correspond,
     NULL_OBJECT