+++ clean-workspace.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+rm -r public
+rm -r posts
+rm -r extra
+rm -r resources
+rm config.json